+++ src/typings/index.ts
@@ -1,0 +1,8 @@
+export type HeartBeator<T> = {
+  poll: Function
+  cancel: Function
+  restart: Function
+  onSuccess(data: T): void
+  onError(err: Error): void
+  onStop: Function
+}
